+++ Sources/KeyboardKit/Layout/InputSet+Rows.swift
@@ -32,17 +32,18 @@ public extension InputSet.Row {
      is mapped to an ``InputSet/Item``.
      */
     init(chars: [String]) {
-        self = chars.map { InputSet.Item($0) }
+        self = chars.map {
+            InputSet.Item($0)
+        }
     }
 
     /**
      Create an input row from a lowercased and an uppercased
      string, where each char is mapped to an ``InputSet/Item``.
 
-     Both arrays must contain the same amount of characters.
+     Both strings must contain the same amount of characters.
      */
-    init(lowercased: String, uppercased: String
-    ) {
+    init(lowercased: String, uppercased: String) {
         self.init(
             lowercased: lowercased.chars,
             uppercased: uppercased.chars
@@ -56,7 +57,8 @@ public extension InputSet.Row {
      Both arrays must contain the same amount of characters.
      */
     init(lowercased: [String], uppercased: [String]) {
-        assert(lowercased.count == uppercased.count, "The lowercased and uppercased string arrays must contain the same amount of characters")
+        let equal = lowercased.count == uppercased.count
+        assert(equal, "lowercased and uppercased must contain the same number of characters")
         self = lowercased.enumerated().map {
             InputSet.Item(
                 neutral: lowercased[$0.offset],
@@ -75,7 +77,7 @@ public extension InputSet.Row {
         pad: String,
         deviceType: DeviceType = .current
     ) {
-        self.init(chars: deviceType == .pad ? pad.chars : phone.chars)
+        self.init(chars: deviceType == .pad ? pad : phone)
     }
 
     /**
@@ -103,7 +105,8 @@ public extension InputSet.Row {
     ) {
         self.init(
             lowercased: deviceType == .pad ? padLowercased.chars : phoneLowercased.chars,
-            uppercased: deviceType == .pad ? padUppercased.chars : phoneUppercased.chars)
+            uppercased: deviceType == .pad ? padUppercased.chars : phoneUppercased.chars
+        )
     }
 
     /**
@@ -119,7 +122,8 @@ public extension InputSet.Row {
     ) {
         self.init(
             lowercased: deviceType == .pad ? padLowercased : phoneLowercased,
-            uppercased: deviceType == .pad ? padUppercased : phoneUppercased)
+            uppercased: deviceType == .pad ? padUppercased : phoneUppercased
+        )
     }
     
     /**